无关大雅的改动，float->double

+++ trunk/app_all/accountBook/ABTransactionDetail.cpp
@@ -62,7 +62,7 @@ void ABTransactionDetail::setInitData(QDate date,
                                      ABCategory category,
                                      QString fromAccount,
                                      QString toAccount,
-                                     float sum,
+                                     double sum,
                                      QString detail)
 {
     ui->dateEdit->setDate(date);
